Upgrade to .NET 8 and update package references; move CookieAuthenticationOptions converter and related classes into separate files.
