+++ EntityFramework/EntityFramework.csproj
@@ -1,16 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders" Version="1.5.0" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders.Saml" Version="1.6.0" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders.WsFederation" Version="1.3.0" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders.EntityFramework" Version="1.4.0" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0.1" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders" Version="2.0.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders.Saml" Version="2.0.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders.WsFederation" Version="2.0.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders.EntityFramework" Version="2.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="8.0.0" />
   </ItemGroup>
 
 </Project>
