+++ IdentityServer/HostingExtensions.cs
@@ -1,0 +1,70 @@
+﻿using IdentityServerHost.Quickstart.UI;
+
+namespace IdentityServer
+{
+    public static class HostingExtensions
+    {
+        public static WebApplication ConfigureServices(this WebApplicationBuilder builder)
+
+        {
+            builder.Services.AddControllersWithViews();
+
+            builder.Services.AddIdentityServer(options =>
+                {
+                    options.Events.RaiseErrorEvents = true;
+                    options.Events.RaiseInformationEvents = true;
+                    options.Events.RaiseFailureEvents = true;
+                    options.Events.RaiseSuccessEvents = true;
+
+                    options.EmitStaticAudienceClaim = true;
+                })
+                .AddTestUsers(TestUsers.Users)
+                .AddInMemoryIdentityResources(Config.IdentityResources)
+                .AddInMemoryApiScopes(Config.ApiScopes)
+                .AddInMemoryClients(Config.Clients)
+                .AddDeveloperSigningCredential();
+
+            builder.Services.AddDynamicProviders(options =>
+               {
+                   // Component setup
+                   options.Licensee = "";
+                   options.LicenseKey = "";
+
+               })
+               .AddJsonStore(options => options.Path = "schemes.json")
+               .AddOpenIdConnect()
+               // .AddSaml(o =>
+               // {
+               //     o.Licensee = "";
+               //     o.LicenseKey = "";
+               // })
+               ;
+            return builder.Build();
+        }
+
+        public static WebApplication ConfigurePipeline(this WebApplication app)
+        {
+            if (!app.Environment.IsDevelopment())
+            {
+                app.UseExceptionHandler("/Home/Error");
+                app.UseHsts();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseDeveloperExceptionPage();
+
+            app.UseStaticFiles();
+            app.UseRouting();
+
+            app.UseIdentityServer();
+            app.UseAuthorization();
+            // Map routes
+            app.MapControllerRoute(
+                name: "default",
+                pattern: "{controller=Home}/{action=Index}/{id?}");
+
+            return app;
+        }
+    }
+}
