+++ IdentityServer/IdentityServer.csproj
@@ -1,15 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="IdentityServer4" Version="4.1.2" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders" Version="1.5.0" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders.Saml" Version="1.6.0" />
-    <PackageReference Include="Rsk.DynamicAuthenticationProviders.WsFederation" Version="1.3.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="4.1.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders" Version="2.0.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders.Saml" Version="2.0.0" />
+    <PackageReference Include="Rsk.DynamicAuthenticationProviders.WsFederation" Version="2.0.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="5.0.0" />
     <PackageReference Include="Rsk.Saml.IdentityServer4" Version="5.3.1" />
   </ItemGroup>
